requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-restframework==0.0.1
djangorestframework==3.14.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
